+++ examples/nextjs/src/config/index.ts
@@ -1,9 +1,7 @@
-import { getWalletConnectConnector, Wallet } from '@rainbow-me/rainbowkit';
-import { Chain, configureChains, mainnet } from 'wagmi';
-import { CoinbaseWalletConnector } from 'wagmi/connectors/coinbaseWallet';
-import { InjectedConnector } from 'wagmi/connectors/injected';
-import { MetaMaskConnector } from 'wagmi/connectors/metaMask';
-import { publicProvider } from 'wagmi/providers/public';
+import { connectorsForWallets } from '@rainbow-me/rainbowkit';
+import { rainbowWallet, walletConnectWallet, trustWallet } from '@rainbow-me/rainbowkit/wallets';
+import { createConfig, http } from 'wagmi';
+import { Chain, mainnet, bscGreenfield } from 'wagmi/chains';
 import * as env from './env';
 
 export const GRPC_URL = env.GRPC_URL;
@@ -25,7 +23,7 @@ export const GROUPHUB_ADDRESS = env.GROUPHUB_ADDRESS;
 
 const greenFieldChain: Chain = {
   id: GREEN_CHAIN_ID,
-  network: 'greenfield',
+  name: 'greenfield',
   rpcUrls: {
     default: {
       http: [GREENFIELD_RPC_URL],
@@ -34,18 +32,16 @@ const greenFieldChain: Chain = {
       http: [GREENFIELD_RPC_URL],
     },
   },
-  name: 'greenfield',
   nativeCurrency: {
-    name: 'BNB',
-    symbol: 'BNB',
+    name: 'tBNB',
+    symbol: 'tBNB',
     decimals: 18,
   },
 };
 
-const bscChain: Chain = {
+export const bscChain: Chain = {
   id: BSC_CHAIN_ID,
   name: 'BSC',
-  network: 'QA - bsc smart chain',
   rpcUrls: {
     default: {
       http: [BSC_RPC_URL],
@@ -65,67 +61,36 @@ const bscChain: Chain = {
   // testnet: true,
 };
 
-const { chains, publicClient, webSocketPublicClient } = configureChains(
+console.log('bscChain', bscChain);
+console.log('greenFieldChain', greenFieldChain);
+
+const connectors = connectorsForWallets(
   [
+    {
+      groupName: 'Recommended',
+      wallets: [rainbowWallet, walletConnectWallet, trustWallet],
+    },
+  ],
+  {
+    appName: 'gnfd demo',
+    projectId: '9bf3510aab08be54d5181a126967ee71',
+  },
+);
+
+export const wagmiConfig = createConfig({
+  chains: [
     mainnet,
+    bscChain,
     {
       ...greenFieldChain,
+      network: 'greenfield',
       iconUrl:
         'https://github.com/wagmi-dev/wagmi/assets/5653652/44446c8c-5c72-4e89-b8eb-3042ef618eed',
     },
-    bscChain,
   ],
-  [publicProvider()],
-);
-
-const coinbaseWalletConnector = new CoinbaseWalletConnector({
-  chains,
-  options: {
-    appName: 'wagmi',
+  transports: {
+    [mainnet.id]: http(),
+    [bscChain.id]: http(),
   },
+  connectors,
 });
-
-const trustWalletConnector = new InjectedConnector({
-  chains,
-  options: {
-    name: 'GN',
-    shimDisconnect: true,
-    // TODO: rainbowkit conflict
-    getProvider: () => (typeof window !== 'undefined' ? (window as any).trustwallet : undefined),
-  },
-});
-
-const metaMaskWalletConnector = new MetaMaskConnector({ chains });
-
-export interface MyWalletOptions {
-  projectId: string;
-  chains: Chain[];
-}
-
-const RainbowTrustWalletConnector = ({ chains, projectId }: MyWalletOptions): Wallet => ({
-  id: '_trust-wallet',
-  name: 'Trust Wallet',
-  iconUrl: 'https://my-image.xyz',
-  iconBackground: '#0c2f78',
-  downloadUrls: {
-    android: 'https://play.google.com/store/apps/details?id=my.wallet',
-    ios: 'https://apps.apple.com/us/app/my-wallet',
-    chrome: 'https://chrome.google.com/webstore/detail/my-wallet',
-    qrCode: 'https://my-wallet/qr',
-  },
-  createConnector: () => {
-    return {
-      connector: trustWalletConnector,
-    };
-  },
-});
-
-export {
-  publicClient,
-  webSocketPublicClient,
-  chains,
-  metaMaskWalletConnector,
-  coinbaseWalletConnector,
-  trustWalletConnector,
-  RainbowTrustWalletConnector,
-};