+++ packages/js-sdk/src/types/common.ts
@@ -10,3 +10,11 @@ export {
 } from '@bnb-chain/greenfield-cosmos-types/greenfield/storage/common';
 export { Long };
 import Long from 'long';
+
+export type OnProgressEvent = {
+  direction: string;
+  percent: number;
+  total: number;
+  loaded: number;
+};
+export type OnProgress = (event: OnProgressEvent) => void;