+++ packages/js-sdk/src/types/sp/DelegatedPubObject.ts
@@ -1,12 +1,14 @@
-import { DelegatedOpts, ResumableOpts } from './Common';
+import { OnProgress } from '../common';
+import { DelegatedOpts, ResumableOpts, UploadFile } from './Common';
 
 export type DelegatedPubObjectRequest = {
   bucketName: string;
   objectName: string;
-  body: File;
+  body: UploadFile;
   delegatedOpts: DelegatedOpts;
   endpoint?: string;
   timeout?: number;
   contentType?: string;
   resumableOpts?: ResumableOpts;
+  onProgress?: OnProgress;
 };